+++ sample-standard-usage/src/main/java/fr/sii/ogham/sample/standard/sms/BasicSampleExternalProperties.java
@@ -1,5 +1,6 @@
 package fr.sii.ogham.sample.standard.sms;
 
+import java.io.IOException;
 import java.util.Properties;
 
 import fr.sii.ogham.core.builder.MessagingBuilder;
@@ -7,17 +8,12 @@ import fr.sii.ogham.core.exception.MessagingException;
 import fr.sii.ogham.core.service.MessagingService;
 import fr.sii.ogham.sms.message.Sms;
 
-public class BasicSample {
+public class BasicSampleExternalProperties {
 
-	public static void main(String[] args) throws MessagingException {
-		// configure properties (could be stored in a properties file or defined
-		// in System properties)
+	public static void main(String[] args) throws MessagingException, IOException {
+		// load properties (available at src/main/resources)
 		Properties properties = new Properties();
-		properties.setProperty("ogham.sms.smpp.host", "<your server host>");
-		properties.setProperty("ogham.sms.smpp.port", "<your server port>");
-		properties.setProperty("ogham.sms.smpp.systemId", "<your server system ID>");
-		properties.setProperty("ogham.sms.smpp.password", "<your server password>");
-		properties.setProperty("ogham.sms.from", "<phone number to display for the sender>");
+		properties.load(BasicSampleExternalProperties.class.getResourceAsStream("/sms.properties"));
 		// Instantiate the messaging service using default behavior and
 		// provided properties
 		MessagingService service = new MessagingBuilder().useAllDefaults(properties).build();